moves sms & email to queue
